fix: restore fixed-precision jiff formatting and add driver encoding assertions
Bring back the fixed 9-digit nanosecond formatting (`{value:.9}`) for
Timestamp, Zoned, Time, and DateTime string casts that was lost in the
revert of 67982d2. This ensures lexicographic string sorting matches
chronological order (ISO 8601 uniform-precision guarantee).

Move the inline unit tests from `stmt/ty_jiff.rs` to a dedicated
`toasty-core/tests/stmt_cast_jiff.rs` integration test file.

Restore driver-level assertions in the jiff integration tests that
verify Timestamp values are encoded correctly (native vs text) when
sent to the database.
