+++ crates/toasty-core/Cargo.toml
@@ -26,4 +26,5 @@ jiff = ["dep:jiff"]
 serde = ["dep:serde", "bit-set/serde"]
 
 [dev-dependencies]
+jiff.workspace = true
 pretty_assertions.workspace = true
